changing atoi for long int insteand of int

+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: yallo <yallo@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/09 14:16:47 by yallo             #+#    #+#             */
-/*   Updated: 2023/08/29 19:55:49 by yallo            ###   ########.fr       */
+/*   Updated: 2023/10/24 14:18:36 by yallo            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@
 # include <unistd.h>
 # include <stdint.h>
 
-int				ft_atoi(const char *str);
+long			ft_atoi(const char *str);
 void			ft_bzero(void *src, size_t n);
 int				ft_isalnum(int c);
 int				ft_isalpha(int c);
